inline parsing funcs and remove redundant TestRecordMetadata test
