+++ destination_test.go
@@ -56,49 +56,9 @@ func TestParsePineconeVector(t *testing.T) {
 	is.Equal(vec.Id, "key1")
 
 	is.Equal(vec.Values, vecToBeWritten.Values)
-	is.Equal(vec.SparseValues, vecToBeWritten.PineconeSparseValues())
+	is.Equal(vec.SparseValues, vecToBeWritten.SparseValues.PineconeSparseValues())
 
 	metadata := vec.Metadata.AsMap()
 	is.Equal(metadata["prop1"], "val1")
 	is.Equal(metadata["prop2"], "val2")
-}
-
-func TestRecordMetadata(t *testing.T) {
-	is := is.New(t)
-
-	var rec sdk.Record
-	rec.Metadata = sdk.Metadata{
-		"created_at": "2023-03-15T14:25:07Z",
-		"prop1":      "a1",
-		"prop2":      "a2",
-	}
-	recMetadata, err := parsePineconeMetadata(rec)
-	is.NoErr(err)
-
-	is.Equal(recMetadata.Fields["prop1"].AsInterface().(string), "a1")
-	is.Equal(recMetadata.Fields["prop2"].AsInterface().(string), "a2")
-}
-
-func testRecords(op sdk.Operation, keys ...string) []sdk.Record {
-	recs := make([]sdk.Record, len(keys))
-	for i := range recs {
-		var position sdk.Position
-		var key sdk.Data = sdk.RawData(keys[i])
-		var metadata sdk.Metadata
-		var payload sdk.Data = sdk.StructuredData{
-			"vector": []float64{1, 2},
-		}
-
-		rec := sdk.Record{
-			Position: position, Operation: op,
-			Metadata: metadata, Key: key,
-			Payload: sdk.Change{
-				Before: nil,
-				After:  payload,
-			},
-		}
-		recs[i] = rec
-	}
-
-	return recs
 }
